expose more build parallelism; renamed target nanogui_python to nanogui-python
