+++ .gitignore
@@ -16,6 +16,7 @@ Makefile
 bin2c
 example1
 example2
+example3
 icons
 *~
 nanogui.so
