+++ python/python.h
@@ -1,6 +1,9 @@
 #if !defined(__PYTHON_H)
 #define __PYTHON_H
 
+#include <nanogui/nanogui.h>
+#include <nanogui/opengl.h>
+
 #include <pybind11/stl.h>
 #include <pybind11/operators.h>
 #include <pybind11/functional.h>
@@ -31,4 +34,11 @@
         NANOGUI_SCREEN_OVERLOADS(Name); \
     }
 
+namespace py = pybind11;
+using namespace nanogui;
+
+/// Make pybind aware of the ref-counted wrapper type
+PYBIND11_DECLARE_HOLDER_TYPE(T, ref<T>);
+
 #endif /* __PYTHON_H */
+
